Add files from brown shoes
This adds the files we need to get started. All specs pass. This commit
adds the lib/white_shoes files because some swt specs depend on them.
The white_shoes files should be removed.

This also adds instructions to the README that should be sufficient to
get people set up and running specs.