Fix issues with backend loading, and move backend-specific behavior to the front.
With this patch it becomes possible to do

````
SHOES_BACKEND=swt rspec spec/shoes/app_spec.rb
````

To run a single spec against a specific backend.

Also, Shoes::Swt::EditBox and EditLine were setting defaults styles on the dsl
object. This is *wrong*. Now these defaults are in Shoes::Edit{Box,Line} so
they are the same on all backends.