Initial commit of relative dimensions support
This will resolve issue #380. The Dimensions class needs an update to draw the
width and/or height from the parent, which now needs to be passed in to work
properly.

For the time being I've chosen to make it graceful if an object doesn't pass
the parent, but this might not be necessary once all the DSL elements properly
support the parent.