Use contents_alignment instead of new overrides
Previously thought that we couldn't make the right behavior with only using the
contents_alignment method, but turns out we can. Before was simply missing the
necessary overrides to let the TextBlock customize all the required absolute_*
methods in question, which originally stalled out my attempts to use the
existing hook.