Bit better treatment of with text spacing
When text blocks had larger :leading values, it revealed that we were removing
too much from the next positioning. Big part of it was where we were
taking out double the leading, which before had looked right, but only at small
leading values.

There are still some differences in the exact placement compared with Shoes 3,
bit this is a distinct improvement for the time being.