Better name for height condition
If we're at the start of a new container, or if the preceding element has used
up all the horizontal space above us, then the y and next_line_start of the
current position will be equal.

This commit doesn't change the logic, but instead gives it a name so it's more
comprehensible.