Text flowing for samples/sample45
samples/sample45.rb had line breaking wrong from flowing text with different
fonts. If a text block fit on a single line but, because of font, was taller
than the preceding height, we bumped down prematurely.

These changes fixes #517, the issue reporting that. sample45 looked wrong on
Shoes 3 and 4.  I've left it for revealing the flow issue, but made
simple-font.rb with better alignment as an alternate, go-forward example.