Refactoring TextBlockPainter to make links easier
The existing TextBlockPainter logic assumed that we only had a single
underlying text layout object to interact with. As such, methods treated
styling as one operation when some apply to all layouts equally, others to only
one of the layouts at a time.

This changeset splits those apart so the defaulting of styles across both
layouts happens in one set of methods, changes that are specific to a given
layout in others.

There was also some renaming to make better sense of what's happening. Names
were a bit generic--coming in it wasn't clear why set_font_styles was different
from set_text_styles, but they are actually dramatically different things. The
new names hopefully convey that more clearly.