Gating merging styles on segment having opts
If we get something other than a Span (or later a Link) then the gathering on
the styles wont have #opts to call against. In this case just skip over it and
apply any other segment styling we can find in the list.

This replaces the class-specific check for Shoes::Span and is a little more
general to the same effect. We will probably still need a class-specific check
for the link styling, but haven't looked deeply at that yet.