Make LinkSegments from FittedTextLayoutCollection
We need a better home for a lot of the link creation code, and it gets even
more complicated with the addition of splits across layouts. Time for a new
class!

Shoes::Swt::LinkSegment which is only stubbed out for now, will be responsible
for estabishing the actual clickable SWT regions that we care about for a link.
A list of these will be tracked on the DSL Link class, although they'll be of
limited use in most cases.

This commit wires up the necessary method in the FittedTextLayoutCollection for
creating these links based on the ranges we've got, with the hard work of
making something actually clickable put off to another commit.