Basic flawed implementation of flowing links
This implements a basic scheme to handle links that flow across multiple
lines of text in a layout. However, I noted almost as I was getting it
working that it assumes the link is starting on the first link, and also
completely misses the mark if multiple links show up across a TextBlock.

Anyway, it's a valid checkpoint along the way with some solid tests
which will still need to pass wherever the implementation for this
heads.