Add #dispose, with implementation for TextBlock
As part of the Common::Clear module, there is now a dispose hook. In the
ordinary case, the method doesn't do anything, but it's now a lifecycle
event. Now, any class that creates Swt resources which must be disposed
can override #dispose and do what they need to do, like TextBlock in
this commit.