Add global keypress for shutdown with leak hunter
If we're running leak hunter, then elements that are still live at the
shutdown of the app will get reported as leaked incorrectly.

This change registers a global keypress for the leak hunter on
Ctrl+Alt+Q that will clear the contents before calling to quit, letting
everything clean up neatly.

It's possible in the future we could alter things to allow hooking the
close events more directly, or by default run the clear, but for now
this helps clear out the noise and also gives us a nice place to wire in
internal diagnostics should we desire more of them on a keypress.