Initial work on clicks passing elements
For all but the application and slots, we actually want click handlers
to pass along the element clicked, not the mouse information (button, x,
y).

This commit does the production work to bake in the necessary parameter
for it to the Shoes::Common::Clickable module, and wires it up for
proper behavior in the SWT backend.

This only adds the spec to TextBlock and leaves the few that had it
before for the shared example "clickable backend". This should
definitely be applied more broadly, but will do that in another commit.