Fix running bin/shoes.bat from source
On gem installation, we copy ext/install/shoes.bat over to the Ruby bin
directory. The batch file assumed that it would be located next to
ruby-shoes to execute.

When running from source, though, we were "linking" bin/shoes.bat to just
point to the ext/install/shoes.bat location. On *nix systems, the symlink
would behave as if it was in the bin directory, but on Windows the file
is literally just a single line reading "ext\install\shoes.bat", which
executes the ext copy of shoes.bat, but from the wrong location.

The fix here was to make ext/install/shoes.bat take a bin_dir if already
set, no longer treat bin/shoes.bat as a link to that file, and have
bin/shoes properly set the bin_dir so running ext/install/shoes.bat works.

Fun times in trying to not have your Ruby program get started by Ruby :)