Cache color for fill and stroke
A large amount of painting time when we had a lot of elements went into
regenerating the fill and stroke colors. These don't actually change
often, so that's wasted time.

Caching these in the SWT layer requires, though, that we get notified
when there's a styling change. Like visibility, this gets method calls
when color updates occur.