Move SWT initialization around
Once ::Swt.display is touched, we can't set the title and have it apply
to the Mac menu bar anymore (it does show in the window title for
whatever reason). So, we need to postpone anything that touches the
display until ::Shoes::Swt::App actually has a chance to get started.

Turns out this is easy enough if we just postpone loading fonts and
system colors. This also simplifies the font loading some, as it was
making this weird transit through the DSL, despite being initiated in
the SWT/backend side of the world. Now we just admit that the backend
must load fonts on its own timeline and move on with our lives.