Factor parent visibility in
When visibility changes happen, we use update_visibility to notify the
elements that might need to directly change their current visibility
state. This was only abiding by the direct visible styling for the
element, so when a parent container got hidden we had to hack around
with tromping visibility at the lower levels.

Instead we now take advantage of our hidden? method and have that factor
in the parent visibility. If direct visibility access is still needed,
it should go through the styling, but that's almost never what you want.

Note that this works without need a new backend element for the slots,
which is good because I'm convinced none of the SWT elements will
actually work since they don't support transparency. :sob: